+++ lib/ui/page/home/introduction/view.dart
@@ -170,8 +170,8 @@ class IntroductionView extends StatelessWidget {
           }
 
           return AnimatedSizeAndFade(
-            fadeDuration: const Duration(milliseconds: 150),
-            sizeDuration: const Duration(milliseconds: 200),
+            fadeDuration: const Duration(milliseconds: 250),
+            sizeDuration: const Duration(milliseconds: 250),
             child: ListView(
               key: Key('${c.displayPassword.value}${c.displaySuccess.value}'),
               shrinkWrap: true,